requirements.txt
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cloudevents==1.11.0
colorama==0.4.6
cryptography==42.0.8
deprecation==2.1.0
Flask
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==3.1.1
functions-framework==3.8.0
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
psycopg2-binary==2.9.6
pycparser==2.22
PyJWT==2.6.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
watchdog==4.0.1
Werkzeug==3.0.3
